backend/uauth/serializers.py
from uauth.models import User
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('url', 'id', 'u_name', 'u_password', 'is_super')